sw content controls: add LOK API
This is somewhat similar to LOK_CALLBACK_FORM_FIELD_BUTTON: if the
cursor enters or leaves a content control, then we send this message, so
the LOK client can render some kind of shading and/or border around the
content control to indicate the boundaries of the object.

Similar to selections, this can be multiple rectangles in case the
string is long enough that the layout breaks it into multiple lines.

Change-Id: I0641a19503b7a1d4cade8fe9b510605cab49f258
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133314
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins