Access2Base - (access2base.py) Don't use bare except: statement
The exception hierarchy described on page
  https://docs.python.org/3.5/library/exceptions.html#exception-hierarchy
indicates that the default exception in bare
  except:
statements is BaseException. This induces that the SystemExit,
KeyboardInterrupt and GeneratorExit are also handled by the
user script.

This is a not recommended practice.

Better is to use the explicit Exception built-in exception
  except Exception:

Bug reported by Paul M on Telegram

Change-Id: Ie1ae1f732ebc60a881e7d40ba8141aa704e9cd5c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133328
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>