SwNavigator: Improve list order for text fields in text frames
Fields in text frames may not be in order of document appearance in
the Navigator Fields list. For example, when a field in the first
paragraph has a greater start position in the paragraph than the
start position of a field in the second paragraph, the field in the
second paragraph is placed before the field in the first paragraph in
the Fields list in the Navigator. This patch fixes this unexpected
behavior by first sorting the fields in document model order and then
doing a second sort using the anchor position of text frames for the
fields in text frames to place these fields in document layout order.

Change-Id: If86f133fbce72f936334dffbc237d097de382ca5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133350
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>