ScriptForge - (SF_Calc) FIX CompactUp/Left whan range is 1 row/col
When the range to compact spans only 1 row (CompactUp)
or 1 column (CompactLeft), a GetValue() method returns
a scalar and not an array as expected in usual cases.

This caused an abort of the script when invoking the
UBound() value of the returned array with a Basic run-time error
   Object variable not set

A test on being scalar or array of the result of
the GetValue() method avoids the error.
To be done in both CompactXXX() methods.

Impact only on Basic code. Python not involved.
No changes due in documentation.

Change-Id: I74cf2b274944b442f843e1f4b7ac3a05dd6eaae7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134352
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>