sw content controls, date: add doc model & UNO API
This is meant to be a content control (providing rich text), which also
has a dropdown-like date picker button.

Add a new Date property to track this type, together with date format
and date language. This should be enough for the UI to generate a
correct date string when the file picker is used.

Change-Id: If5d46a804d771e903a688fd73cfaf2d2809b7ab9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134847
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins