sw content controls, date: preserve more properties
While working on the DOCX import for dates, it turns out there is a need
to store more properties for DOCX export purposes. Given that these are
potentially useful for full support, add dedicated UNO API and DOCX
export for these (i.e. not just grab-bag), but omit UI or ODT filter for
now. This includes:

- <w:docPart w:val="...">

- <w:dataBinding w:prefixMappings="..." w:xpath="..." w:storeItemID="...">

- <w15:color w:val="...">

And tests for all these.

Change-Id: I18ddec50d40c1c4abd87f7ea947a24dd8a92a755
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135039
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins