Fix comment
For one, the functions are already available in C++20.  For another, our
fallback implementation is only an approximation, as it doesn't enforce the
requirement that the involved template types "are standard integer types or
extended integer types" (but which is at least checked in --with-latest-c++
builds against standard libraries that already provide those C++20 functions).

Change-Id: I91208c185f2db681cf25899d3cd9a9e2912d32a0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135232
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>