remove incorrect #ifdef DBG_UTIL around normal code (tdf#149476)
This code sets sufficient size for the virtual device that saves
the content, so it's not some optional debug code. Without this
selecting text in Calc formula bar and dragging it around leaves
artifacts, because the saved content is not large enough to restore
the original widget content. This dates back to
d51fe1189dcbe0fe06805c8d99687c125cd254eb, which supposedly made
the code warning-free (and also correct-free).

Change-Id: I52dab8c8d3a083fff9da8aed27facdf21076622e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135555
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>