tdf#148868 sw: handle selection then Insert similar to Replace
... if the selection is inside one paragraph, to avoid deleting flys
anchored in the selection.

From a code point of view it's a bit inconsistent to do this, but
from user point of view there are some ways to conveniently create
a selection such as by double clicking a word.

(see also tdf#133957)

Also in SwWrtShell::AutoCorrect(), which oddly enough might get called
with a selection from textsh*.cxx, at least in theory.

Change-Id: I8cf9459e5a7ec7754ce8fe323cd158c7e84a5c93
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135606
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>